Make nREPL implementation more generic
- Add checks for supported nREPL operations where missing.

- Check for a "runtime" field on the "versions" field returned by the
"describe" operation. If supported and the runtime is not "clj", disable
features that rely on Clojure.

- Support associating major modes with a runtime, using the
`cider-repl-types-for-buffer` variable.

- Replace `cider-repl-init-code` with
`cider-repl-require-repl-utils-code`, where the correct init code will
be run based on the REPL's runtime.

- When running `cider-read-and-eval`, use the major mode of the active
buffer instead of assuming `clojure-mode`.

- Modify REPL startup message based on runtime type.

- Add interactive function `cider-connect-generic` for connecting to a
non-Clojure nREPL. In this case the runtime will be inferred based on
the "runtime" field provided by the nREPL server.

- The nREPL port prompt now also picks up projects without a Clojure
build tools file but that have a .nrepl-port file or similar.